improve memory level statistics

+++ src/mem_analyzer.h
@@ -8,16 +8,16 @@
 typedef uint64_t date_t;
 
 struct count {
-  unsigned count;
+  uint64_t count;
   uint64_t min_weight;
   uint64_t max_weight;
   uint64_t sum_weight;
 };
 
 struct mem_counters {
-  unsigned total_count;
-  unsigned total_weight;
-  unsigned na_miss_count;
+  uint64_t total_count;
+  uint64_t total_weight;
+  uint64_t na_miss_count;
 
   struct count cache1_hit;
   struct count cache2_hit;
